+++ packages/fds-mui-theme/src/overrides/MuiIconButton.js
@@ -1,0 +1,9 @@
+const MuiIconButton = {
+  root: {
+    '&:hover': {
+      backgroundColor: 'transparent',
+    },
+  },
+};
+
+export default MuiIconButton;