+++ CHANGELOG.md
@@ -17,6 +17,36 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [1.13]
+
+#### **Changes**
+- Added `fds-mui-theme` package. This provides a theme for `material-ui` based on our standard styles
+  defined in `fds-dictionary`.
+
+
+## [1.11, 1.12]
+bad publish - identical to `1.10`.
+
+## [1.10]
+
+#### **Changes**
+- Added material UI palettes to `fds-dictionary`. Light and dark variants are generated for each
+  base color for use in `material-ui` framework theming.
+
+```
+import with: `import { blue } from 'fds-dictionary/dist/js/materialPalette.js'`
+
+blue[500]; // base blue color
+blue[100]; // lightened blue
+blue[900]; // darkened blue
+```
+
+## [1.9]
+
+#### **Changes**
+- Added distribution to `fds-dictionary`, a map of CSS custom property names to values. This is used
+  to populate a PostCSS plugin with fallback values of custom properties for IE11.
+
 ## [1.8]
 
 #### **Changes**